extract pm quality score routes
